+++ packages/ethereum-contracts/CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - EVM target changed from _shanghai_ to _cancun_.
+- Update solhint to v6.
 
 ## [v1.14.1]