+++ packages/ethereum-contracts/CHANGELOG.md
@@ -3,11 +3,12 @@ All notable changes to the ethereum-contracts will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [v1.15.0]
 
 ### Added
 
 - `SuperToken`: the contract admin can enable/disable a _Yield Backend_ in order to generate a yield on the underlying asset.
+- `SuperToken`: added versioning. `VERSION()` now returns the version string of the logic contract set for the SuperToken.
 
 ### Changed